+++ src/main/java/kr/or/kosa/backend/pay/entity/PointHistory.java
@@ -1,6 +1,8 @@
 package kr.or.kosa.backend.pay.entity;
 
 import lombok.*;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @AllArgsConstructor
@@ -13,7 +15,7 @@ public class PointHistory {
 
     private Long id;               // AUTO_INCREMENT PK
     private String userId;
-    private int changeAmount;      // + 적립, - 사용
+    private BigDecimal changeAmount;      // + 적립, - 사용
     private String type;           // EARN / USE / REFUND 등
     private String paymentOrderId; // 어느 결제와 묶였는지 (nullable)
     private String description;    // 설명
